[refactor]CardsetContainer에서 화면 그리는 부분 컴포넌트로 분리 - isViewMoreHidden을 아예 다 삭제하고 클릭 상태를 'child.id === clickedCardsetId'로 판단하도록 수정 - CardsetContainer에서 UI 그려주는 부분 다 Component(CardsetBox & CardBox)로 분리
